replace app-layout and update routes
