+++ apps/app/src/routes/home-old/home.page.tsx
@@ -4,6 +4,7 @@ import { useAuth } from 'react-oidc-context';
 import { Link } from 'react-router-dom';
 import { ReactFlowProvider } from 'reactflow';
 import PositionProfile from '../../components/app/common/components/positionProfile';
+import ContentWrapper from '../../components/content-wrapper.component';
 import { useGetPositionRequestsCountQuery } from '../../redux/services/graphql-api/position-request.api';
 import { useGetProfileQuery } from '../../redux/services/graphql-api/profile.api';
 import MyPositionsTable from '../my-position-requests/components/my-position-requests-table.component';
@@ -18,7 +19,6 @@ import { OrgChartType } from '../org-chart/enums/org-chart-type.enum';
 import HeaderWrapper from './components/header-wrapper.component';
 import { InitialsAvatar } from './components/initials-avatar.component';
 import './home.page.css';
-import ContentWrapper from '../../components/content-wrapper.component';
 
 const { Text } = Typography;
 
@@ -114,7 +114,7 @@ export const HomePage = () => {
             itemsPerPage={5}
             tableTitle="My recent positions"
             topRightComponent={
-              <Link to="/my-position-requests">
+              <Link to="/requests/positions">
                 <Button type="link" style={{ padding: '0' }}>
                   View all positions
                 </Button>
