+++ apps/app/src/routes/home/components/default-home-page/components/org-chart-card.component.tsx
@@ -42,7 +42,7 @@ export const OrgChartCard = () => {
         }
         extra={
           <Space direction="horizontal">
-            <Link to="/next/my-departments">
+            <Link to="/my-departments">
               <Button type="link" style={{ padding: '0' }}>
                 View all organizations
               </Button>
