+++ apps/app/src/routes/job-profiles/components/job-profile-search-results.component.tsx
@@ -36,7 +36,7 @@ export const JobProfileSearchResults = ({
   const [searchParams] = useSearchParams();
   const location = useLocation();
 
-  const isPositionRequestRoute = location.pathname.includes('/my-position-requests/');
+  const isPositionRequestRoute = location.pathname.includes('/requests/positions/');
   const { positionRequestId } = useParams<{ positionRequestId?: string }>();
 
   const getBasePath = (path: string) => {
@@ -55,21 +55,19 @@ export const JobProfileSearchResults = ({
     //   searchParams.toString().length > 0 ? `?${searchParams.toString()}` : ''
     // }`
 
-    const next = location.pathname.includes('next') ? '/next' : '';
-
-    // Check if we're on /saved-profiles route
-    if (location.pathname.includes('/saved-profiles')) {
-      return `${next}/saved-profiles/${profileNumber}`;
+    // Check if we're on /job-profiles/saved route
+    if (location.pathname.includes('/job-profiles/saved')) {
+      return `/job-profiles/saved/${profileNumber}`;
     }
 
-    // Check if we're on the position-request route
+    // Check if we're on the /requests/positions route
     const newSearchParams = new URLSearchParams(searchParams.toString());
     if (positionRequestId) {
       newSearchParams.set('selectedProfile', profileNumber.toString());
-      return `${next}/my-position-requests/${positionRequestId}?${newSearchParams.toString()}`;
+      return `/requests/positions/${positionRequestId}?${newSearchParams.toString()}`;
     } else {
-      // If not on the position-request route, use the standard job-profiles path
-      return `${next}/job-profiles/${profileNumber}?${newSearchParams.toString()}`;
+      // If not on the /requests/positions route, use the standard job-profiles path
+      return `/job-profiles/${profileNumber}?${newSearchParams.toString()}`;
     }
   };
