+++ apps/app/src/routes/job-profiles/components/job-profiles.component.tsx
@@ -121,7 +121,7 @@ const JobProfiles = forwardRef<JobProfilesRef, JobProfilesContentProps>(
 
     const getBasePath = useCallback(
       (path: string) => {
-        if (positionRequestId) return `/my-position-requests/${positionRequestId}`;
+        if (positionRequestId) return `/requests/positions/${positionRequestId}`;
 
         const pathParts = path.split('/');
         // Check if the last part is a number (ID), if so, remove it
