+++ apps/app/src/routes/total-comp-draft-profiles/components/total-comp-draft-profiles-table.component.tsx
@@ -131,16 +131,16 @@ const TotalCompProfilesTable: React.FC<MyPositionsTableProps> = ({
     if (is_archived === false) {
       // eslint-disable-next-line react-hooks/rules-of-hooks
       [trigger, { data, isLoading, error: fetchError }] = useLazyGetJobProfilesDraftsQuery();
-      link = '/draft-job-profiles/';
+      link = '/job-profiles/manage/draft/';
     } else {
       // eslint-disable-next-line react-hooks/rules-of-hooks
       [trigger, { data, isLoading, error: fetchError }] = useLazyGetJobProfilesArchivedQuery();
-      link = '/archived-job-profiles/';
+      link = '/job-profiles/manage/archived/';
     }
   } else {
     // eslint-disable-next-line react-hooks/rules-of-hooks
     [trigger, { data, isLoading, error: fetchError }] = useLazyGetJobProfilesQuery();
-    link = '/published-job-profiles/';
+    link = '/job-profiles/manage/published/';
   }
 
   // Check if data is available and call the callback function to notify the parent component
@@ -185,8 +185,6 @@ const TotalCompProfilesTable: React.FC<MyPositionsTableProps> = ({
     }
   }, [data]);
 
-  const next = location.pathname.includes('next') ? 'next' : '';
-
   const columns: ColumnTypes[] = [
     {
       title: 'Job Title',
@@ -194,11 +192,7 @@ const TotalCompProfilesTable: React.FC<MyPositionsTableProps> = ({
       key: 'title',
       sorter: allowSorting,
       defaultSortOrder: getSortOrder('title'),
-      render: (text: any, record: any) => (
-        <Link to={`${next.length > 0 ? '/next/job-profiles/manage/draft/' : link}${record.id}`}>
-          {text?.trim() || 'Untitled'}
-        </Link>
-      ),
+      render: (text: any, record: any) => <Link to={`${link}${record.id}`}>{text?.trim() || 'Untitled'}</Link>,
     },
     {
       sorter: allowSorting,
@@ -774,7 +768,7 @@ const TotalCompProfilesTable: React.FC<MyPositionsTableProps> = ({
               <>
                 <div>Looks like you’re not working on anything right now.</div>
                 {/* Link button to the orgchart page */}
-                <Link to="/draft-job-profiles/create">
+                <Link to="/job-profils/manage/drafts/create">
                   <Button type="primary" style={{ marginTop: '1rem' }}>
                     Create new profile
                   </Button>
